requirements.txt
watchdog>=0.9.0
pyyaml>=3.0

# test requirements
pytest
pytest-cov